% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{arrange_row}
\alias{arrange_row}
\title{Sort a data frame}
\usage{
arrange_row(df, ...)
}
\arguments{
\item{df}{a data frame.}

\item{...}{further parameters to \code{order}.}
}
\value{
a sorted data frame.
}
\description{
sort a date fram and keep the row names.
}
\details{
This function is modified from \code{arrange} of R package \code{plyr}.
\code{arrange} deliberately remove rownames, but sometimes it is worth to
keep them.
}
