% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{batch_shift}
\alias{batch_shift}
\title{Batch shifting}
\usage{
batch_shift(x, y, method = "mean")
}
\arguments{
\item{x}{a data matrix.}

\item{y}{a categorical data for batch/block information.}

\item{method}{method for shifting.}
}
\value{
returns a shifted data matrix.
}
\description{
Remove batch effect withing each block.
}
\references{
Silvia Wagner, et.al, Tools in Metabonomics: An Integrated Validation
Approach for LC-MS Metabolic Profiling of Mercapturic Acids in Human
Urine Anal. Chem., 2007, 79 (7), pp 2918-2926, DOI: 10.1021/ac062153w
}
