% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{bi_cor_net}
\alias{bi_cor_net}
\title{Bipartite/two-mode correlation network}
\usage{
bi_cor_net(co_mat, thres = 0.6, dn = NULL)
}
\arguments{
\item{co_mat}{a correlation coffeficient matrix.}

\item{thres}{correlation coeeficient threshold for network analysis.}

\item{dn}{a character string for correlation name.}
}
\value{
a list of \code{ggplot2} plots.
}
\description{
Perform bipartite/two-mode correlation network analysis.
}
