% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{blank_filter}
\alias{blank_filter}
\title{MS/NMR data filtering based on "blank" data}
\usage{
blank_filter(
  x,
  y,
  method = c("mean", "median", "max"),
  factor = 1,
  f_mv = TRUE,
  thres_mv = 0.3
)
}
\arguments{
\item{x}{a data matrix.}

\item{y}{a character string with contents of "sample", "qc" and "blank".}

\item{method}{method for stats. Support "mean", "median" and "max".}

\item{factor}{multiplier for blank stats}

\item{f_mv}{a flag indicating whether perform missing value filtering on
"sample" data.}

\item{thres_mv}{threshold of missing values on QC. Features less than this
threshold will be kept.}
}
\value{
a list of with contents: \itemize{
\item dat the filtered data matrix
\item idx a logical vector of index for keeping features.
}
}
\description{
MS/NMR data filtering based on "blank" data
}
\details{
This function provides an option to perform missing value filtering
on "sample" data.
}
\seealso{
Other variable filters: 
\code{\link{locfdr_filter}()},
\code{\link{mv_filter}()},
\code{\link{qc_filter}()},
\code{\link{rsd_filter}()},
\code{\link{var_filter}()}
}
\concept{variable filters}
