% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{chunk}
\alias{chunk}
\title{Split a vector into chunks}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{an vector.}

\item{n}{length of chunk.}
}
\value{
return the split chunk.
}
\description{
Split a vector into chunks.
}
\details{
For details, see https://bit.ly/2SM4m2G
}
\examples{
x <- 1:10
n <- 3
chunk(x, n)
}
