% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{cor_hcl}
\alias{cor_hcl}
\title{Correlation hierarchical cluster analysis}
\usage{
cor_hcl(
  mat,
  cutoff = 0.75,
  use = "pairwise.complete.obs",
  method = "pearson",
  fig_title = "Cluster Dendrogram",
  size = 3,
  rotate = FALSE
)
}
\arguments{
\item{mat}{a data matrix}

\item{cutoff}{a threshold for correlation analysis.}

\item{use}{a string giving a method for computing covariances. For details,
see \code{cor}.}

\item{method}{method for correlation \code{cor}.}

\item{fig_title}{title for plotting}

\item{size}{text font size of \code{ggplot2}.}

\item{rotate}{a logical indicating whether to rotate plot.}
}
\value{
a list of a \code{ggplot2} object for clustering and a list of
cluster centres.
}
\description{
Perform hierarchical cluster analysis based on correlation.
}
\examples{
cor_hcl(mtcars)
}
