% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{cor_net}
\alias{cor_net}
\title{Correlation network analysis}
\usage{
cor_net(
  mat,
  use = "pairwise.complete.obs",
  method = "pearson",
  thres = 0.6,
  fig_title = "Correlation network"
)
}
\arguments{
\item{mat}{a data matrix for correlation analysis}

\item{use, method}{parameters of \code{cor}.}

\item{thres}{correlation coeeficient threshold for network analysis. Only
keep those with coeeficient larger than \code{thres}.}

\item{fig_title}{a chracter string for gifure title.}
}
\value{
a list of \code{ggplot2} plots.
}
\description{
Perform network analysis for correlation coefficient.
}
