% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{cor_tab}
\alias{cor_tab}
\title{Caculate correlation coefficient and p-values}
\usage{
cor_tab(
  x,
  cor.method = c("pearson", "kendall", "spearman"),
  adj.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"),
  ...
)
}
\arguments{
\item{x}{a data frame or matrix for correlation analysis column-wise.}

\item{cor.method}{method for correlation}

\item{adj.method}{p-value correction method}

\item{...}{other parameter for correlation.}
}
\value{
a list of with contents: \itemize{
\item r correlation coefficient
\item p statistics matrix, in which the lower triangular is p-values
and the upper triangular is adjusted p-values
}
}
\description{
Calculate correlation coefficient and p-values.
}
\details{
This file is modified from 'cor.table' of package 'picante'
and 'corr.test' of package 'psych'.
The original implementation is from Bill Venables, the author of R great
book MASS. For details, see
https://stat.ethz.ch/pipermail/r-help/2001-November/016201.html
}
\examples{
co <- cor_tab(mtcars, cor.method = "spearman", adj.method = "BH")
names(co)

\dontrun{
library(pysch)
co <- corr.test(mtcars, method="spearman",adjust="BH")
# From pysch: For symmetric matrices, p values adjusted for multiple tests
# are reported above the diagonal.
}
}
