% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{dat2long}
\alias{dat2long}
\title{Melt a numeric data matrix to long format}
\usage{
dat2long(x)
}
\arguments{
\item{x}{a matrix or data frame}
}
\value{
returns a \code{tibble} object
}
\description{
Reshape a matrix or data frame to long format with row names and colomn
names as two columns.
}
\details{
\code{reshape2::melt} keeps the rownames when melting a matrix, but
not when melting a data frame. This function keeps rownames for both
matrix and data frame.
}
\examples{
class(mtcars)     # data.frame
reshape2::melt(mtcars)
reshape2::melt(as.matrix(mtcars))
dat2long(mtcars)
}
