% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{dat_cor}
\alias{dat_cor}
\title{Correlation between two data sets}
\usage{
dat_cor(
  dat_1,
  dat_2,
  partial = FALSE,
  method = c("pearson", "kendall", "spearman"),
  ...
)
}
\arguments{
\item{dat_1, dat_2}{two data matrix for correlation analysis. Note that they
must have the same row numbers.}

\item{partial}{a logical value indicating whether to perform partial
correlation analysis.}

\item{method}{correlation method.}

\item{...}{further parameters to correlation method.}
}
\value{
a mstrix of correlation.
}
\description{
Performs correlation between two data sets
}
\details{
This function is very specific. Will change for more general
purpose.
}
