% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{dat_filter}
\alias{dat_filter}
\title{MS/NMR data filtering}
\usage{
dat_filter(x, method = "var_filter", ...)
}
\arguments{
\item{x}{a list of metabolomics data, including intensity data and peak
information.}

\item{method}{the method for filtering.}

\item{...}{further parameters to be passed to `method`.}
}
\value{
a filtered list of metabolomics data.
}
\description{
Wrapper function for filtering MS/NMR data.
}
