% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{dat_summ}
\alias{dat_summ}
\title{Data matrix summary}
\usage{
dat_summ(x, method = mean, ...)
}
\arguments{
\item{x}{a matrix-like object.}

\item{method}{summary method for an vector.}

\item{...}{further parameters for \code{method}.}
}
\value{
retuns a summarised table.
}
\description{
Summarise a data matrix.
}
\examples{
\dontrun{
iris \%>\% group_by(Species) \%>\% group_modify(~ df_summ(., method = mean))
iris \%>\% group_by(Species) \%>\% group_modify(~ df_summ(., method = vec_stats))
iris \%>\% group_by(Species) \%>\% do(df_summ(., method = vec_segment))
iris \%>\% df_summ(method = sd)
}
}
