% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{dat_symb}
\alias{dat_symb}
\title{Symbolise data set}
\usage{
dat_symb(x, thres)
}
\arguments{
\item{x}{an vector or data matrix.}

\item{thres}{an vector with lower and higher values.}
}
\value{
a symbolic data set.
}
\description{
Symbolise data values as (-1, 0, 1) based on a range
}
\details{
\code{thres} can be estimated by \code{\link[=locfdr_filter]{locfdr_filter()}}.
}
\seealso{
\code{\link[=locfdr_filter]{locfdr_filter()}} and \code{\link[=locfdr]{locfdr()}} for \code{z.2}.
}
