% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{df_na_idx}
\alias{df_na_idx}
\title{Get row index of missing values in data frame}
\usage{
df_na_idx(df, vars)
}
\arguments{
\item{df}{data frame being used to check index of missing values.}

\item{vars}{vector of variable names in \code{df} for missing values checking.}
}
\value{
returns an vector of missing value index.
}
\description{
Get row index of missing values in data frame.
}
