% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{df_t}
\alias{df_t}
\title{Transpose a numeric data frame (with/without NAs)}
\usage{
df_t(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a transposed data frame.
}
\description{
Transpose a numeric data frame (with/without NAs)
}
\details{
It is used mostly for numeric data frame with NAs and return a
}
\examples{
 
x <- data.frame(group = c(1,1,2,NA,2), val = c(6,4,6,3.1,NA))
x <- df_t(x)
}
