% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{dim_mat}
\alias{dim_mat}
\title{Get dimension matrix of a list of data frame}
\usage{
dim_mat(mat_list)
}
\arguments{
\item{mat_list}{a list of data frame.}
}
\value{
a data frame.
}
\description{
Retuns a matrix for dimension information of a list of data frame.
}
