% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{feat_count}
\alias{feat_count}
\title{Caculate the count number of features}
\usage{
feat_count(fs.ord, top.k = 30)
}
\arguments{
\item{fs.ord}{data matrix for multiple feature order lists.}

\item{top.k}{top feature number to be processed.}
}
\value{
a data matrix of feature counts.
}
\description{
Calculate the features counts based on the frequency of multiple selectors.
}
