% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{gg_heat_dend}
\alias{gg_heat_dend}
\title{Heatmap with dendrograms with ggplot2}
\usage{
gg_heat_dend(
  mat,
  row.dend = T,
  col.dend = T,
  row.dend.right = TRUE,
  colors = c("red", "white", "blue"),
  font.size = 10,
  x.rot = 60,
  legend.title = "value",
  dist.method = "euclidean",
  clust.method = "complete",
  dend.line.size = 0.5
)
}
\arguments{
\item{mat}{a data mstrix to be plotted.}

\item{row.dend}{plot row dendrogram or not.}

\item{col.dend}{plot column dendrogram ot not.}

\item{row.dend.right}{a ligical value to indivate the position of row
dendrogram.}

\item{colors}{a vector of colours for heatmap.}

\item{font.size}{label font size.}

\item{x.rot}{plot rotate degree.}

\item{legend.title}{lengend title.}

\item{dist.method}{distance method.}

\item{clust.method}{cluster method.}

\item{dend.line.size}{dendrogram line size.}
}
\value{
retuns an object of class \code{ggplot2}.
}
\description{
Plot heatmap of a data matrix using \code{ggplot2}. This funcion is modofied
from https://bit.ly/2UUnY2L.
}
\examples{
gg_heat_dend(mtcars)
}
