% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{graph_stats}
\alias{graph_stats}
\title{Graph level metrics}
\usage{
graph_stats(g)
}
\arguments{
\item{g}{an \code{igraph} object.}
}
\value{
a vector of graph level metrics.
}
\description{
Return graph level metrics.
}
