% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{heat_dend}
\alias{heat_dend}
\title{Heatmap with dendrogram on both sides}
\usage{
heat_dend(
  mat,
  x.rot = 60,
  col = c("red", "white", "blue"),
  cex = 0.5,
  xlab = "",
  ylab = "",
  main = "",
  ...
)
}
\arguments{
\item{mat}{a data matrix for plotting.}

\item{x.rot}{the rotate degree.}

\item{col}{colours for heatmap.}

\item{cex}{font size.}

\item{xlab, ylab, main}{character strings for \code{xlab}, \code{ylab} and figure title.}

\item{...}{further parameters for \code{lattice}.}
}
\value{
an object of class \code{lattice}.
}
\description{
Plot heatmap with dendrogram on both sides using lattice
}
\examples{
heat_dend(mtcars)
}
