% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{locfdr_filter}
\alias{locfdr_filter}
\title{Filtering variable based on local false discovery rate}
\usage{
locfdr_filter(x, plot = 1, thres = NULL, ...)
}
\arguments{
\item{x}{a data metirx}

\item{plot}{an interger for plotting. 0 gives no plots. 1 gives single
plot showing the histogram of zz and fitted densities f and p0*f0.}

\item{thres}{a user defined threshold for filtering. The default is NULL,
which use local FDR as threshold for filtering.}

\item{...}{other parameters to be passed to \code{locfdr}.}
}
\value{
a list of with contents: \itemize{
\item dat the filtered data matrix.
\item idx a vector of filtering index.
\item thres threshold used for filtering.
}
}
\description{
Filter data based on local false discovery rate. This function uses \code{z.2}
of \code{locfdr}.
}
\details{
From R package \code{locfdr} vignette:
\code{z.2} is the interval along the zz-axis outside of which \code{fdr(z) < 0.2},
the locations of the yellow triangles in the histogram plot. If no
elements of \code{zz} on the left or right satisfy the criterion, the
corresponding element of z.2 is NA, and the corresponding triangle does
not appear.
}
\seealso{
\code{\link[=locfdr]{locfdr()}}
}
