% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{mat2df}
\alias{mat2df}
\title{Convert matrix to df}
\usage{
mat2df(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
returns a data frame.
}
\description{
Convert matrix to df
}
\details{
from \code{.matrix_to_df} of package \code{plyr}.
}
