% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{mm_scale}
\alias{mm_scale}
\title{Scale vector between min and max}
\usage{
mm_scale(x, low, high)
}
\arguments{
\item{x}{an vector.}

\item{low}{lower value.}

\item{high}{upper value.}
}
\value{
returns scaled vector.
}
\description{
Scale vector between mininum and maxinum values.
}
\details{
From \code{rescale} in package \code{network}
}
