% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\docType{package}
\name{mtExtra-package}
\alias{mtExtra}
\alias{mtExtra-package}
\title{mtExtra: More functions for metabolomics data analysis}
\description{
\code{mtExtra} provides some general functions and more specific
functions to R package \code{mt} for metabolomics data analysis.
}
\section{Main functions}{

The main mtExtra provides more functions for metabolomics data analysis.
These functions are statistical analysis and plot methods with \code{ggplot2}
and \code{lattice}. It uses \code{tidyverse} as well as \code{reshape2} and \code{plyr}
packages.
}

\section{Package context}{

This package follows the principles of the "tidyverse" as mush as possible.
It also uses \code{melt} in \code{reshape2} if \code{tidyr} is complicated in some
circumstances.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wanchanglin/mtExtra}
  \item Report bugs at \url{https://github.com/wanchanglin/mtExtra/issues}
}

}
\keyword{internal}
