% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\docType{package}
\name{mtExtra}
\alias{mtExtra}
\title{mtExtra}
\description{
`mtExtra` has more functions for metabolomics data analysis. It should use
with R package \code{mt}.
}
\section{Main functions}{

The main mtExtra provides more functions for metabolomics data analysis.
These functions are statistical analysis and plot methods with `ggplot2` and
`lattice`. It uses `tidyverse` as well as `reshape2` and `plyr` packages.
}

\section{Package context}{

This package follows the principles of the "tidyverse" as mush as possible.
It also uses `melt` in `reshape2` if `tidyr` is complicated in some
circumstances.
}

\keyword{internal}
