% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{mv_filter}
\alias{mv_filter}
\title{Filtering features based on the percentage of missing values}
\usage{
mv_filter(x, thres = 0.3)
}
\arguments{
\item{x}{an vector, matrix or data frame.}

\item{thres}{a threshold to drop off the variables.}
}
\value{
a list of with contents: \itemize{
\item x the filtered data matrix
\item idx a vector of filtering index.
}
}
\description{
This function calculates the percentage of missing values and keeps those
features with missing values percentage less than the designed threshold.
}
