% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{mv_perc}
\alias{mv_perc}
\title{Missing value percentage}
\usage{
mv_perc(x)
}
\arguments{
\item{x}{an vector, matrix or data frame.}
}
\value{
missing value percentage.
}
\description{
Calculate missing value percentage.
}
\examples{
meta <- mtExtra:::meta
mv_perc(meta)
mv_filter(meta, thres = 0.3)
}
\seealso{
\code{\link[=mv_filter]{mv_filter()}}
}
