% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{net_graph}
\alias{net_graph}
\title{Create an igraph object}
\usage{
net_graph(edge_df, node_df = NULL)
}
\arguments{
\item{edge_df}{graph eddge matrix.}

\item{node_df}{graph vertex matrix.}
}
\value{
an \code{igraph} object with some vertex statistics and community
detection center.
}
\description{
Create an \code{igraph} object
}
