% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{non_digit}
\alias{non_digit}
\title{Calculate the percentage of non digits}
\usage{
non_digit(mat)
}
\arguments{
\item{mat}{a data matrix.}
}
\value{
returns an vector.
}
\description{
Calculate the percentage of non digits.
}
