% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{non_zero}
\alias{non_zero}
\title{Calculate the non-zero percentage}
\usage{
non_zero(mat)
}
\arguments{
\item{mat}{a data matrix.}
}
\value{
returns an vector.
}
\description{
Calculate the non-zero percentage.
}
