% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA plot}
\usage{
pca_plot(x, y = NULL, scale = TRUE, ep.plot = FALSE, ...)
}
\arguments{
\item{x}{a matrix for plotting. Should have row names}

\item{y}{a factor or character vector specifiing the group information of
the row.}

\item{scale}{a logical flag to indicate whether or not scale \code{x}.}

\item{ep.plot}{plot ellipse or not.}
}
\value{
a base R graphics object.
}
\description{
Plot PCA of a matrix or data frame with base R function
}
\examples{
data(iris)
pca_plot(iris[, 1:4], iris[, 5], ep.plot = TRUE)
}
\keyword{multivariate}
