% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{plot_aam}
\alias{plot_aam}
\title{Wrapper function for plotting classification restults}
\usage{
plot_aam(aam, fig_title = "Accuracy, AUC and Margin")
}
\arguments{
\item{aam}{a data matrix of classification results.}

\item{fig_title}{a string of figure title}
}
\value{
an object of class `ggplot2`
}
\description{
This function plots accuracy, AUC and margin (aam) of classification 
results from package \code{mt}.
}
