% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{plot_pval}
\alias{plot_pval}
\title{Plot adjusted p-values}
\usage{
plot_pval(pval_list)
}
\arguments{
\item{pval_list}{a data matrix or a list of data matrix of p-value correction}
}
\value{
an object of class \code{ggplot2}
}
\description{
Plot the adjusted p-values using \code{ggplot2}
}
\examples{
pval <- mtExtra:::pval
plot_pval(pval)
}
