% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{qc_filter}
\alias{qc_filter}
\title{MS/NMR data filtering based on RSD of "qc" data}
\usage{
qc_filter(
  x,
  y,
  thres_rsd = 20,
  f_mv = TRUE,
  f_mv_qc_sam = FALSE,
  thres_mv = 0.3
)
}
\arguments{
\item{x}{a data matrix.}

\item{y}{a character string with contents of "sample", "qc" and "blank".}

\item{thres_rsd}{threshold of RSD on QC. Features less than this
threshold will be kept.}

\item{f_mv}{a flag indicating whether or not to performance missing value
filtering on either "sample" or "qc" data.}

\item{f_mv_qc_sam}{a flag for filtering using percentage of missing values
on "qc" or "sample". TRUE is for "qc".}

\item{thres_mv}{threshold of missing values. Features less than this
threshold will be kept.}
}
\value{
a filtered data.
}
\description{
MS/NMR data filtering based on RSD of "qc" data
}
\details{
This filter process takes two steps. First, the missing values
filterting is performed on either "qc" or "sample". Then RSD-based
filtering is applied to "qc" data.
}
\seealso{
Other variable filters: 
\code{\link{blank_filter}()},
\code{\link{locfdr_filter}()},
\code{\link{mv_filter}()},
\code{\link{rsd_filter}()},
\code{\link{var_filter}()}
}
\concept{variable filters}
