% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_qc.R
\name{qc_pass}
\alias{qc_pass}
\title{Data filtering based on quality assurance in QC samples}
\usage{
qc_pass(x, y, tol = 0.3, thres = 0.4)
}
\arguments{
\item{x}{A data matrix.}

\item{y}{A character string with contents of "sample", "qc" and "blank".}

\item{tol}{A torrance for quality assurance}

\item{thres}{A threshold for quality assurance}
}
\value{
a list of with contents: \itemize{
\item dat the filtered data matrix
\item idx a logical vector of index for keeping features.
}
}
\description{
Data filtering based on quality assurance in QC samples
}
