% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{qc_rlsc}
\alias{qc_rlsc}
\title{Quality control–based robust LOESS signal correction (QC-RLSC)}
\usage{
qc_rlsc(x, y, ...)
}
\arguments{
\item{x}{a data matrix.}

\item{y}{a categorical data for batch/block information.}

\item{...}{other parameter for 'loess' such as 'span' (the parameter
which controls the degree of smoothing. Default: 0.75) and
'degree' (the degree of the polynomials to be used. Default: 2).}
}
\value{
a corrected data matrix.
}
\description{
Signal correction based on quality control.
}
\references{
Warwick B Dunn, et.al, Procedures for large-scale metabolic profiling of
serum and plasma using gas chromatography and liquid chromatography
coupled to mass spectrometry, Nature Protocols, 6:1060–1083 (2011)
}
