% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{rbind_df}
\alias{rbind_df}
\title{Row binding of a list of data matrix}
\usage{
rbind_df(list)
}
\arguments{
\item{list}{a list of data matrix.}
}
\value{
returns a data mstrix.
}
\description{
Row bind of a list of matrix or data frame with the same dimension.
}
