% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{rsd_filter}
\alias{rsd_filter}
\title{Filtering variable based on relative standard deviation (RSD)}
\usage{
rsd_filter(x, thres = 20)
}
\arguments{
\item{x}{a data frame where columns are features.}

\item{thres}{threshold of RSD. Features less than this threshold will be
kept.}
}
\value{
a list of with contents: \itemize{
\item dat the filtered data matrix
\item idx a logical vector of index for keeping features.
}
}
\description{
Filtering variable based on relative standard deviation (RSD).
}
\seealso{
Other variable filters: 
\code{\link{blank_filter}()},
\code{\link{locfdr_filter}()},
\code{\link{mv_filter}()},
\code{\link{qc_filter}()},
\code{\link{var_filter}()}
}
\concept{variable filters}
