% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{samp_strat}
\alias{samp_strat}
\title{Select random sample with stratification}
\usage{
samp_strat(grp, len, strat = TRUE)
}
\arguments{
\item{grp}{a character string indicating two group information.}

\item{len}{length of random selection.}

\item{strat}{a logical value indicated the sampling should be stratified.}
}
\value{
returns a selected index.
}
\description{
Select random sample with stratification from a binary group.
}
\details{
multiple group is not supported at the current stage.
}
\examples{
cls <- iris[, 5, drop = TRUE]
cls <- cls[cls == "setosa" | cls == "versicolor"]
cls <- droplevels(cls)
ind <- samp_strat(cls, 6, strat = TRUE)
cls[ind]
ind <- samp_strat(cls, 5, strat = TRUE)
cls[ind]
}
