% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{samp_sub}
\alias{samp_sub}
\title{Select random samples from each group

Select random samples from each group.}
\usage{
samp_sub(x, k, n = 10)
}
\arguments{
\item{x}{a character string or factor indicating class info.}

\item{k}{number of samples selected from each class.}

\item{n}{number of replicates.}
}
\value{
returns a selected index.
}
\description{
Select random samples from each group

Select random samples from each group.
}
\examples{
 
 cls   <- iris[, 5, drop = TRUE]
 (tmp <- samp_sub(cls,k = 6, n = 2))
 cls[tmp[[1]]]
 table(cls[tmp[[1]]])
}
