% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{sd_filter}
\alias{sd_filter}
\title{Filtering variable based SD}
\usage{
sd_filter(x, sig.cutoff = 0.95, na.rm = FALSE)
}
\arguments{
\item{x}{an vector, matrix or data frame.}

\item{sig.cutoff}{a numeric value (default 0.95) for `quantile`.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
a list of with contents: \itemize{
 \item x the filtered data matrix
 \item idx a vector of filtering index.
}
}
\description{
This function calculates the standard deviation for every feature across all
samples and keep the features with large quantile controlled by `sig.cutoff`.
}
