% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{sd_zero_filter}
\alias{sd_zero_filter}
\title{Filtering variables with sd values close to zero}
\usage{
sd_zero_filter(x, na.rm = FALSE)
}
\arguments{
\item{x}{an vector, matrix or data frame.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
a list of with contents: \itemize{
\item x the filtered data matrix
\item idx a vector of filtering index.
}
}
\description{
Filter variables with sd values close to zero
}
