% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{std_scale}
\alias{std_scale}
\title{Scale vector with standarization}
\usage{
std_scale(x, na.rm = TRUE)
}
\arguments{
\item{x}{an vector.}

\item{na.rm}{a logical for whether removing NAs.}
}
\value{
returns scaled vector
}
\description{
Perform standarization of an vector.
}
