% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{str_trim}
\alias{str_trim}
\title{Trim white spaces}
\usage{
str_trim(string)
}
\arguments{
\item{string}{a character string to be processed.}
}
\value{
returns a trimmed string.
}
\description{
Trim white spaces.
}
