% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{dat_trans}
\alias{dat_trans}
\alias{vec_trans}
\alias{trans}
\title{Transform data}
\usage{
dat_trans(x, method = "auto", add = 1)

vec_trans(x, method = "auto", add = 1)
}
\arguments{
\item{x}{a matrix, data frame or vector.}

\item{method}{transformation method, including: "center", "auto", "range",
"pareto", "vast", "level", "log", "log10", "sqrt" and "asinh".}

\item{add}{a shift value for log transformation.}
}
\value{
transformed data
}
\description{
Perform data set transformation
}
\examples{
 data(iris)
 ## transform an vector
 vec <- iris[, 1]
 dat_trans(vec, method = "auto")
 vec_trans(vec, method = "auto")
 ## transform a data frame
 mat <- iris[, 1:4]
 dat_trans(mat, method = "log")
 ## transform data frame under condition
 plyr::ddply(iris, ("Species"), function(x, method) {
   dat_trans(x[, 1:4], method = method)
 }, method = "range")

## use 'tidyverse'
library(dplyr)
library(tidyr)
library(purrr)

## transform whole data set
iris \%>\% mutate(across(where(is.numeric), ~ vec_trans(., method = "range")))
## transform data set with group
iris \%>\% 
  group_by(Species) \%>\%
  mutate(across(where(is.numeric), ~ vec_trans(., method = "range")))
}
\references{
Berg, R., Hoefsloot, H., Westerhuis, J., Smilde, A. and Werf, M. (2006),
Centering, scaling, and transformations: improving the biological
information content of metabolomics data, \emph{BMC Genomics}, 7:142
}
