% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{trim_str}
\alias{trim_str}
\title{Trim white spaces}
\usage{
trim_str(string)
}
\arguments{
\item{string}{a character string to be processed.}
}
\value{
a trimmed string.
}
\description{
Trim head and hail white spaces of a string.
}
