% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_2.R
\name{upd_data}
\alias{upd_data}
\title{Update data set by a subset of features}
\usage{
upd_data(dat.all, ord)
}
\arguments{
\item{dat.all}{a list of metabolomics data including intensity data, peak
data and meta data}

\item{ord}{a subset of selected feature orders}
}
\description{
Update data set after feature selection.
}
