% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{var_filter}
\alias{var_filter}
\title{Filtering variable based on variability}
\usage{
var_filter(x, method = "IQR", na.rm = FALSE, thres = 0.25)
}
\arguments{
\item{x}{a matrix or data frame.}

\item{method}{variability measurement method, such as \code{IQR} and \code{sd}.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{thres}{a numeric value between 0 and 1 for the threshold of
\code{quantile}. Features whose variability value is large than this threshold
will be kept.}
}
\value{
a list of with contents: \itemize{
\item dat the filtered data matrix
\item idx a vector of filtering index.
}
}
\description{
Perform variable filtering based on feature variability measurement such as
standard deviation(SD) and inter-quantile range(IQR).
}
\examples{
set.seed(100)
x <- matrix(rnorm(20 * 10), ncol = 10)
res <- var_filter(x, method = "sd", thres = 0.25)
sum(res$idx)
}
\seealso{
Other variable filters: 
\code{\link{blank_filter}()},
\code{\link{locfdr_filter}()},
\code{\link{mv_filter}()},
\code{\link{qc_filter}()},
\code{\link{rsd_filter}()}
}
\concept{variable filters}
