% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{vec_ci}
\alias{vec_ci}
\title{Vector summary}
\usage{
vec_ci(x, na.rm = FALSE, conf.interval = 0.95)
}
\arguments{
\item{x}{a numveric vector.}

\item{na.rm}{remove NA or not.}

\item{conf.interval}{a numeric value for confidential interval.}
}
\value{
retuns a vector of summary.
}
\description{
Calculate the statistical summary of a vector, including confidential
interval.
}
\details{
Used for error bar plotting. Modify from
https://bit.ly/3onsqot
}
