% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{vec_norm}
\alias{vec_norm}
\title{vector normalisation}
\usage{
vec_norm(x, method = "median", scale = TRUE)
}
\arguments{
\item{x}{a vector}

\item{method}{the normalisation method to be used.}

\item{cale}{a logical to scale `x` or not}
}
\value{
returns a normalised vector
}
\description{
Normalise a numeric vector
}
\examples{
x <- c(2, 3, 4, 5, 6, 7, 8, 9, 50, 50)
vec_norm(x, method = "median", scale = T)
}
