% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{vec_segment}
\alias{vec_segment}
\title{Vector statistics for error bar plotting}
\usage{
vec_segment(x, bar = c("SD", "SE", "CI"))
}
\arguments{
\item{x}{an vector.}

\item{bar}{a chracter string, supporting "SD", "SE" and "CI".}
}
\value{
retuns an vector including lower, center and upper values.
}
\description{
Calculate vector's standard derivation, standard error of mean and
confidencd interval.
}
\examples{
vec_segment(iris[,1])
mat <- reshape2::melt(iris)
plyr::ddply(mat, plyr::.(Species,variable), function(x,bar) {
  vec_segment(x$value, bar = bar)
}, bar = "SD")
}
\seealso{
Other vector stats functions: 
\code{\link{vec_stats}()}
}
\concept{vector stats functions}
