% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{vec_stats}
\alias{vec_stats}
\title{Vector summary}
\usage{
vec_stats(x, na.rm = FALSE, conf.interval = 0.95)
}
\arguments{
\item{x}{a numveric vector.}

\item{na.rm}{remove NA or not.}

\item{conf.interval}{a numeric value for confidenice interval.}
}
\value{
retuns a vector of summary consisting:
\itemize{
\item number of vector length
\item vector mean
\item vector standard derivation
\item standard error of mean
\item confidence interval
}
}
\description{
Calculate the statistical summary of a vector.
}
\details{
Can be used for error bar plotting. Modify from
https://bit.ly/3onsqot
}
\examples{
 
library(dplyr)
library(tidyr)
library(purrr)

iris \%>\% dat_summ(method = vec_stats)
iris \%>\% group_by(Species) \%>\% group_modify(~ dat_summ(., method = vec_stats))
iris \%>\%
  pivot_longer(cols = !Species, names_to = "var") \%>\% 
  group_nest(Species, var) \%>\%
  mutate(map_dfr(data, ~ vec_stats(.x$value))) \%>\%
  select(!data)
}
\seealso{
Other vector stats functions: 
\code{\link{vec_segment}()}
}
\concept{vector stats functions}
