% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{bi_cor_net}
\alias{bi_cor_net}
\title{Bipartite/two-mode correlation network}
\usage{
bi_cor_net(co_mat, thres = 0.6, dn = NULL)
}
\arguments{
\item{co_mat}{a correlation coefficient matrix.}

\item{thres}{correlation coefficient threshold for network analysis.}

\item{dn}{a character string for correlation name.}
}
\value{
a list of \code{ggplot2} plots.
}
\description{
Perform bipartite/two-mode correlation network analysis.
}
\examples{
data <- mtExtra:::data
meta <- mtExtra:::meta

\dontrun{
## process meta
library(mt)
meta <- mv_filter(meta, thres = 0.3)$dat
meta <- mv.fill(meta, method = "mean") 
meta <- preproc(meta, method = "auto")
}

co <- cor(meta, data, use = "pairwise.complete.obs")
heat_dend(co)

res <- bi_cor_net(co, thres = 0.2)
names(res)
res$net1
res$dend2
}
