% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{cor_hcl}
\alias{cor_hcl}
\title{Correlation hierarchical cluster analysis}
\usage{
cor_hcl(
  mat,
  cutoff = 0.75,
  use = "pairwise.complete.obs",
  method = "pearson",
  fig_title = "Cluster Dendrogram",
  size = 3,
  rotate = FALSE
)
}
\arguments{
\item{mat}{a data matrix}

\item{cutoff}{a threshold for correlation anaalysis.}

\item{use}{a string giving a method for computing covariances. For details, 
see `cor`.}

\item{method}{method for correlation `cor`.}

\item{fig_title}{title for plotting}

\item{size}{text font size of `ggplot2`.}

\item{rotate}{a logical indicating wheter to rotate teh plot.}
}
\value{
retuns a list of a `ggplot2` object for clustering and a list of 
  cluster centres.
}
\description{
Perform hierarchical cluster analysis based on correlation.
}
