% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{cor_net}
\alias{cor_net}
\title{Correlation network analysis}
\usage{
cor_net(
  mat,
  use = "pairwise.complete.obs",
  method = "pearson",
  thres = 0.6,
  fig_title = "Correlation network"
)
}
\arguments{
\item{mat}{a data matrix for correlation analysis}

\item{use, method}{parameters of \code{\link[stats:cor]{stats::cor()}}.}

\item{thres}{correlation coefficient threshold for network analysis. Only
keep those with coefficient larger than \code{thres}.}

\item{fig_title}{a character string for figure title.}
}
\value{
a list of \code{ggplot2} plots.
}
\description{
Perform network analysis for correlation coefficient.
}
\examples{
res <- cor_net(mtcars)
names(res)
res$p1    # or res$p2, res$p3, res$p4
\dontrun{
## see https://ggplot2-book.org/arranging-plots.html
library(patchwork)
res$p1 + res$p2 + res$p3 + res$p4 
}
}
