% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\docType{package}
\name{mtExtra}
\alias{mtExtra}
\alias{mtExtra-package}
\title{Extra functions for metabolomics data analysis}
\description{
\code{mtExtra} has more functions for metabolomics data analysis.
It should use ' with R package \code{mt}.
}
\section{Main functions}{

The main mtExtra provides more functions for metabolomics data analysis.
These functions are statistical analysis and plot methods with \code{ggplot2} and
\code{lattice}. It uses \code{tidyverse} as well as \code{reshape2} and \code{plyr} packages.
}

\section{Package context}{

This package follows the principles of the "tidyverse" as mush as possible.
It also uses \code{melt} in \code{reshape2} if \code{tidyr} is complicated in some
circumstances.
}

