% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA plot}
\usage{
pca_plot(x, y = NULL, scale = TRUE, ep.plot = FALSE, ...)
}
\arguments{
\item{x}{a matrix for plotting. Should have row names}

\item{y}{a factor or character vector specifying the group information of
the row.}

\item{scale}{a logical flag to indicate whether or not scale \code{x}.}

\item{ep.plot}{plot ellipse or not.}

\item{...}{further parameters passed to \code{plot}}
}
\value{
a base R graphics object.
}
\description{
Plot PCA of a matrix or data frame with base R function.
}
\details{
The \code{rownames} of \code{x} will be shown in the plot. \code{y} is used to indicate
the group of row of \code{x}.
}
\examples{
data(iris)
pca_plot(iris[, 1:4], iris[, 5], ep.plot = TRUE)
}
\author{
Wanchang Lin
}
\keyword{multivariate}
