% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{pcor_dat}
\alias{pcor_dat}
\title{Partial correlation analysis}
\usage{
pcor_dat(x, y, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x, y}{two data matrix for correlation analyses.}

\item{method}{correlation method.}
}
\value{
a correlation matrix.
}
\description{
Perform partial correlation analysis of two data matrix
}
