% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{range_scale}
\alias{range_scale}
\title{Scale vector to a defined range}
\usage{
range_scale(x, range = c(0, 1))
}
\arguments{
\item{x}{a numeric vector}

\item{range}{a vector with two values: lower and higher.}
}
\value{
a scaled vector.
}
\description{
Scale vector to a defined range.
}
\examples{
set.seed(100)
x <- rnorm(10)
range_scale(x, range = c(10, 20))
}
\seealso{
Other numeric vector scalers: 
\code{\link{std_scale}()}
}
\concept{numeric vector scalers}
