% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{rsd}
\alias{rsd}
\title{Relative standard deviation}
\usage{
rsd(x, na.rm = TRUE)
}
\arguments{
\item{x}{an vector, matrix or data frame.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
RSD values
}
\description{
Perform Relative Standard Deviation (RSD).
}
\examples{
data(iris)
rsd(iris[, 1:4])
}
