% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra.R
\name{sym2long}
\alias{sym2long}
\title{Convert a symmetric table(short format) to long format}
\usage{
sym2long(x, tri = c("upper", "lower"))
}
\arguments{
\item{x}{a symmetric matrix-like data set.}

\item{tri}{triangular being used.}
}
\value{
returns a data frame of pair-wise comparison.
}
\description{
Convert a symmetric table(short format) to long format
}
\examples{
co <- cor_tab(mtcars, cor.method = "spearman", adj.method = "BH")
names(co)
corr <- sym2long(co$r, tri = "upper")
pval <- sym2long(t(co$p), tri = "upper")
padj <- sym2long(co$p, tri = "upper")
tmp  <- data.frame(corr, pval, padj)
}
