% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{var_filter}
\alias{var_filter}
\title{IQR-based Variable filtering}
\usage{
var_filter(x, var.func = "IQR", var.cutoff = 0.25, filterByQuantile = TRUE)
}
\arguments{
\item{x}{a matrix for processing.}

\item{var.func}{a method for filtering}

\item{var.cutoff}{a threshold for cut-off}

\item{filterByQuantile}{a binary value to control filter by quantitle.}
}
\value{
a list of with contents: \itemize{
 \item x the filtered data matrix
 \item idx a vector of filtring index.
}
}
\description{
Variable filtering(row-wise) based on IQR.
}
\examples{
x <- matrix(rnorm(20*3), ncol = 3)
var_filter(x)
}
