% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extra_1.R
\name{vec_segment}
\alias{vec_segment}
\title{Vector statistics for error bar plotting}
\usage{
vec_segment(x, bar = c("SD", "SE", "CI"))
}
\arguments{
\item{x}{an vector.}

\item{bar}{a character string, supporting "SD", "SE" and "CI".}
}
\value{
an vector including lower, center and upper values.
}
\description{
Calculate vector's standard derivation, standard error of mean and
confidence interval.
}
\examples{
library(plyr)
library(reshape2)
library(dplyr)
library(tidyr)
library(purrr)

vec_segment(iris[,1])

## Use 'plyr' and 'reshape2' for group 
mat <- melt(iris)
ddply(mat, .(Species,variable), function(x,bar) {
  vec_segment(x$value, bar = bar)
}, bar = "SD")

## Use 'tidyverse' for group 
iris \%>\%
  pivot_longer(cols = !Species, names_to = "variable") \%>\% 
  group_by(Species, variable) \%>\%
  nest() \%>\%
  mutate(map_dfr(.x = data, .f = ~ vec_segment(.x$value))) \%>\%
  select(!data)

iris \%>\%
  pivot_longer(cols = !Species, names_to = "variable") \%>\% 
  group_nest(Species, variable) \%>\%
  mutate(map_dfr(data, ~ vec_segment(.x$value))) \%>\%
  select(!data)
}
\seealso{
Other vector stats functions: 
\code{\link{vec_stats}()}
}
\concept{vector stats functions}
